requirements/pip.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
wheel==0.38.4
    # via -r python-template/placeholder_repo_name_0/requirements/pip.in

# The following packages are considered to be unsafe in a requirements file:
pip==23.0.1
    # via -r python-template/placeholder_repo_name_0/requirements/pip.in
setuptools==67.3.2
    # via -r python-template/placeholder_repo_name_0/requirements/pip.in
djangorestframework
markdown
django-filter
django-oauth-toolkit